fix: possible to buyout and extend customerItem

+++ src/app/item/item-display/item-display.component.ts
@@ -78,7 +78,7 @@ export class ItemDisplayComponent implements OnInit {
 		this._userOrderService.alreadyHaveOrderedItem(itemId).then((haveOrdered: boolean) => {
 			this.alreadyOrdered = haveOrdered;
 
-			if (haveOrdered) {
+			if (haveOrdered && !this._cartService.isCustomerItem(itemId)) {
 				this._cartService.remove(itemId); // should remove itself if it is apart of cart for some reason
 			}
 		}).catch((err) => {
@@ -87,7 +87,7 @@ export class ItemDisplayComponent implements OnInit {
 
 		this._userCustomerItemService.alreadyHaveItem(itemId).then((haveItem: boolean) => {
 			this.alreadyHaveItem = haveItem;
-			if (haveItem) {
+			if (haveItem && !this._cartService.isCustomerItem(itemId)) {
 				this._cartService.remove(itemId); // should remove itself if it is apart of cart for some reason
 			}
 		}).catch((err) => {
